Update main.js
Added a new addCommand that can quickly generate column wrapper of 1 without having to manually enter the number of columns the user needs.
Modified the addCommand of the insert-column to have the condition of keeping the user's text when they have them selected. Also modified the cursor location after the command is being performed.